fix(lib): address SonarCloud critical and major code issues
- S5488: Add SAFE_SHUTDOWN macro to guard against invalid fd in client.c
- S5314: Use MSG_DONTWAIT on recv() to prevent blocking in critical sections
- S1066: Merge collapsible if statements in decoder_v4l2.c
- S859: Remove const from vsl_frame_paddr() since it caches paddr internally
- S1905: Remove redundant cast in vsl_timestamp_us()

Updates API signature for vsl_frame_paddr() from const to mutable pointer
since the function modifies internal cached state.

+++ lib/common.c
@@ -23,7 +23,8 @@ vsl_timestamp_us(void)
 {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return (int64_t) ts.tv_sec * 1000000LL + ts.tv_nsec / 1000;
+    // LL suffix ensures 64-bit arithmetic, no cast needed (S1905)
+    return ts.tv_sec * 1000000LL + ts.tv_nsec / 1000;
 }
 
 int
